Moved some tasks of Project read/save
