+++ src/auto_flutter_dig/task/_list.py
@@ -11,8 +11,8 @@ from .flutter.doctor import FlutterDoctor
 from .flutter.generator import FlutterGeneratorTask
 from .flutter.pub.get import FlutterPubGet
 from .project.init import ProjectInit
-from .project.read import ProjectRead
-from .project.save import ProjectSave
+from ..module.aflutter.task.project.read import ProjectRead
+from ..module.aflutter.task.project.save import ProjectSave
 
 task_list: Dict[TaskId, TaskIdentity] = dict(
     sorted(
