+++ src/auto_flutter_dig/task/project/__init__.py
@@ -1,2 +1,2 @@
-from .read import Project, ProjectRead
-from .save import ProjectSave
+from ...module.aflutter.task.project.read import Project, ProjectRead
+from ...module.aflutter.task.project.save import ProjectSave
